requirements.txt
opencv-python
pytesseract
gTTS
python-dotenv
flask
google-cloud-vision
google-generativeai
gtts
httpx
Pillow
